Bug 1460026 - refactor into APIBuilder -> API
