+++ package.json
@@ -46,5 +46,5 @@
   "engines": {
     "node": ">=8"
   },
-  "main": "./src/api.js"
+  "main": "./src/builder.js"
 }
