+++ test/publish_test.js
@@ -1,5 +1,5 @@
 suite('api/publish', function() {
-  var subject         = require('../');
+  var APIBuilder      = require('../');
   var config          = require('typed-env-config');
   var aws             = require('aws-sdk');
   var assert          = require('assert');
@@ -17,7 +17,7 @@ suite('api/publish', function() {
   // Test simple method
   test('publish minimal reference', async function() {
     // Create test api
-    var api = new subject({
+    var builder = new APIBuilder({
       title:        'Test Api',
       description:  'Another test api',
       name:         'test',
@@ -25,92 +25,93 @@ suite('api/publish', function() {
     });
 
     // Declare a simple method
-    api.declare({
+    builder.declare({
       method:       'get',
       route:        '/test0',
       name:         'test0',
       title:        'Test End-Point',
       description:  'Place we can call to test something',
-      stability:    subject.stability.stable,
+      stability:    APIBuilder.stability.stable,
     }, function(req, res) {
       res.send(200, 'Hello World');
     });
 
     // Declare some methods with some fun scopes
-    api.declare({
+    builder.declare({
       method:       'get',
       route:        '/test1',
       name:         'test1',
       title:        'Test End-Point',
       description:  'Place we can call to test something',
-      stability:    subject.stability.stable,
+      stability:    APIBuilder.stability.stable,
       scopes:       {AllOf: ['foo:bar']},
     }, function(req, res) {
       res.send(200, 'Hello World');
     });
-    api.declare({
+    builder.declare({
       method:       'get',
       route:        '/test2',
       name:         'test2',
       title:        'Test End-Point',
       description:  'Place we can call to test something',
-      stability:    subject.stability.stable,
+      stability:    APIBuilder.stability.stable,
       scopes:       {AnyOf: ['foo:bar']},
     }, function(req, res) {
       res.send(200, 'Hello World');
     });
-    api.declare({
+    builder.declare({
       method:       'get',
       route:        '/test3',
       name:         'test3',
       title:        'Test End-Point',
       description:  'Place we can call to test something',
-      stability:    subject.stability.stable,
+      stability:    APIBuilder.stability.stable,
       scopes:       {if: 'not_public', then: {AllOf: ['foo:bar']}},
     }, function(req, res) {
       res.send(200, 'Hello World');
     });
-    api.declare({
+    builder.declare({
       method:       'get',
       route:        '/test4',
       name:         'test4',
       title:        'Test End-Point',
       description:  'Place we can call to test something',
-      stability:    subject.stability.stable,
+      stability:    APIBuilder.stability.stable,
       scopes:       {AllOf: [{for: 'foo', in: 'whatever', each: 'bar:<foo>'}]},
     }, function(req, res) {
       res.send(200, 'Hello World');
     });
-    api.declare({
+    builder.declare({
       method:       'get',
       route:        '/test5',
       name:         'test5',
       title:        'Test End-Point',
       description:  'Place we can call to test something',
-      stability:    subject.stability.stable,
+      stability:    APIBuilder.stability.stable,
       scopes:       {AllOf: [{for: 'foo', in: 'whatever', each: 'bar:<foo>'}]},
     }, function(req, res) {
       res.send(200, 'Hello World');
     });
-    api.declare({
+    builder.declare({
       method:       'get',
       route:        '/test6',
       name:         'test6',
       title:        'Test End-Point',
       description:  'Place we can call to test something',
-      stability:    subject.stability.stable,
+      stability:    APIBuilder.stability.stable,
       scopes:       {if: 'not_public', then: {AllOf: ['abc', {AnyOf: ['e']}, {for: 'a', in: 'b', each: 'c'}]}},
     }, function(req, res) {
       res.send(200, 'Hello World');
     });
 
-    const svc = await api.setup({
+    const api = await builder.build({
       rootUrl:              urls.testRootUrl(),
       referenceBucket:      cfg.referenceTestBucket,
       aws:                  cfg.aws,
     });
 
-    await svc.publish();
+    await api.publish();
+
     // Get the file... we don't bother checking the contents this is good
     // enough
     const s3 = new aws.S3(cfg.aws);
