+++ test/scopes_test.js
@@ -1,12 +1,12 @@
 suite('api/route', function() {
   var assert          = require('assert');
   var Promise         = require('promise');
-  var subject         = require('../');
+  var APIBuilder      = require('../');
   var express         = require('express');
   var path            = require('path');
 
   // Create test api
-  var api = new subject({
+  var builder = new APIBuilder({
     title:        'Test Api',
     description:  'Another test api',
     name:         'test',
@@ -15,7 +15,7 @@ suite('api/route', function() {
 
   test('no scopes is OK', function() {
     // doesn't throw
-    api.declare({
+    builder.declare({
       method:       'get',
       route:        '/test/:myparam',
       name:         'noScopeOktestEP',
@@ -25,7 +25,7 @@ suite('api/route', function() {
   });
 
   test('string scope works', function() {
-    api.declare({
+    builder.declare({
       method:       'get',
       route:        '/testString/:myparam',
       scopes:       'test:unit',
@@ -37,7 +37,7 @@ suite('api/route', function() {
 
   test('array of string scope rejected', function() {
     assert.throws(function() {
-      api.declare({
+      builder.declare({
         method:       'get',
         route:        '/testArr/:myparam',
         scopes:       ['test:unit'],
@@ -50,7 +50,7 @@ suite('api/route', function() {
 
   test('array of arrays of scope rejected', function() {
     assert.throws(function() {
-      api.declare({
+      builder.declare({
         method:       'get',
         route:        '/testArrArr/:myparam',
         scopes:       [[]],
@@ -62,7 +62,7 @@ suite('api/route', function() {
   });
 
   test('scope expression not rejected', function() {
-    api.declare({
+    builder.declare({
       method:       'get',
       route:        '/testScope/:myparam',
       scopes:       {AnyOf: ['something']},
@@ -73,7 +73,7 @@ suite('api/route', function() {
   });
 
   test('scope expression with looping template not rejected', function() {
-    api.declare({
+    builder.declare({
       method:       'get',
       route:        '/testScope2/:myparam',
       scopes:       {AnyOf: [{for: 'foo', in: 'bar', each: '<foo>'}]},
