+++ test/validate_test.js
@@ -2,12 +2,15 @@ suite('api/validate', function() {
   var request         = require('superagent');
   var assert          = require('assert');
   var Promise         = require('promise');
-  var subject         = require('../');
+  var APIBuilder      = require('../');
   var helper          = require('./helper');
   var testing         = require('taskcluster-lib-testing');
+  var libUrls         = require('taskcluster-lib-urls');
+
+  const u = path => libUrls.api(helper.rootUrl, 'test', 'v1', path);
 
   // Create test api
-  var api = new subject({
+  var builder = new APIBuilder({
     title:        'Test Api',
     description:  'Another test api',
     name:         'test',
@@ -15,7 +18,7 @@ suite('api/validate', function() {
   });
 
   // Declare a method we can test input with
-  api.declare({
+  builder.declare({
     method:   'post',
     route:    '/test-input',
     name:     'testInputValidate',
@@ -27,7 +30,7 @@ suite('api/validate', function() {
   });
 
   // Declare a method we can use to test valid output
-  api.declare({
+  builder.declare({
     method:   'get',
     route:    '/test-output',
     name:     'testInputValidOutputValidate',
@@ -39,7 +42,7 @@ suite('api/validate', function() {
   });
 
   // Declare a method we can use to test invalid output
-  api.declare({
+  builder.declare({
     method:   'get',
     route:    '/test-invalid-output',
     name:     'testInputInvalidOutputValidate',
@@ -51,7 +54,7 @@ suite('api/validate', function() {
   });
 
   // Declare a method we can test input validation skipping on
-  api.declare({
+  builder.declare({
     method:   'post',
     route:    '/test-skip-input-validation',
     name:     'testInputSkipInputValidation',
@@ -64,7 +67,7 @@ suite('api/validate', function() {
   });
 
   // Declare a method we can test output validation skipping on
-  api.declare({
+  builder.declare({
     method:   'get',
     route:    '/test-skip-output-validation',
     name:     'testOutputSkipOutputValidation',
@@ -77,7 +80,7 @@ suite('api/validate', function() {
   });
 
   // Declare a method we can test blob output on
-  api.declare({
+  builder.declare({
     method:   'get',
     route:    '/test-blob-output',
     name:     'testBlobOutput',
@@ -89,12 +92,12 @@ suite('api/validate', function() {
   });
 
   // Create a mock authentication server
-  setup(() => helper.setupServer({api}));
+  setup(() => helper.setupServer({builder}));
   teardown(helper.teardownServer);
 
   // Test valid input
   test('input (valid)', function() {
-    var url = 'http://localhost:23525/test-input';
+    var url = u('/test-input');
     return request
       .post(url)
       .send({value: 5})
@@ -106,7 +109,7 @@ suite('api/validate', function() {
 
   // Test invalid input
   test('input (invalid)', function() {
-    var url = 'http://localhost:23525/test-input';
+    var url = u('/test-input');
     return request
       .post(url)
       .send({value: 11})
@@ -118,7 +121,7 @@ suite('api/validate', function() {
 
   // Test valid output
   test('output (valid)', function() {
-    var url = 'http://localhost:23525/test-output';
+    var url = u('/test-output');
     return request
       .get(url)
       .then(function(res) {
@@ -129,7 +132,7 @@ suite('api/validate', function() {
 
   // test invalid output
   test('output (invalid)', function() {
-    var url = 'http://localhost:23525/test-invalid-output';
+    var url = u('/test-invalid-output');
     return request
       .get(url)
       .then(res => assert(false, 'should have failed!'))
@@ -140,7 +143,7 @@ suite('api/validate', function() {
 
   // test skipping input validation
   test('skip input validation', function() {
-    var url = 'http://localhost:23525/test-skip-input-validation';
+    var url = u('/test-skip-input-validation');
     return request
       .post(url)
       .send({value: 100})
@@ -152,7 +155,7 @@ suite('api/validate', function() {
 
   // test skipping output validation
   test('skip output validation', function() {
-    var url = 'http://localhost:23525/test-skip-output-validation';
+    var url = u('/test-skip-output-validation');
     return request
       .get(url)
       .then(function(res) {
@@ -163,7 +166,7 @@ suite('api/validate', function() {
 
   // test blob output
   test('blob output', function() {
-    var url = 'http://localhost:23525/test-blob-output';
+    var url = u('/test-blob-output');
     return request
       .get(url)
       .then(function(res) {
@@ -173,7 +176,7 @@ suite('api/validate', function() {
   });
 
   test('input (correct content-type)', function() {
-    var url = 'http://localhost:23525/test-input';
+    var url = u('/test-input');
     return request
       .post(url)
       .send(JSON.stringify({value: 5}))
@@ -184,7 +187,7 @@ suite('api/validate', function() {
   });
 
   test('input (wrong content-type)', function() {
-    var url = 'http://localhost:23525/test-input';
+    var url = u('/test-input');
     return request
       .post(url)
       .send(JSON.stringify({value: 5}))
